test: move mock response types out of the Docker container module
ExchangeResponse / ResponseBody (and their builder) lived in the Docker/nginx
container module, so the reusable mock-response dataset could not compile
without that harness. Move them to a standalone `response` module and repoint
all mock_responses imports there. This lets the PocketIC harness consume the
same dataset, and lets the Docker harness be removed later without touching the
mocks. No behavioural change.
