+++ src/xrc-tests/src/lib.rs
@@ -8,6 +8,8 @@ mod container;
 #[cfg(test)]
 mod mock_responses;
 #[cfg(test)]
+mod response;
+#[cfg(test)]
 mod templates;
 #[cfg(test)]
 mod tests;
