+++ src/xrc-tests/src/mock_responses/forex.rs
@@ -2,7 +2,7 @@ use std::collections::HashMap;
 
 use xrc::{Forex, FOREX_SOURCES};
 
-use crate::{container::ExchangeResponse, ONE_DAY_SECONDS};
+use crate::{response::ExchangeResponse, ONE_DAY_SECONDS};
 
 mod australia;
 mod bosnia;
